Something PT2 GPT -> Dennis

+++ src/color.h
@@ -3,8 +3,8 @@
 #include <cstdint>
 
 struct Color {
-    uint8_t r;
-    uint8_t g;
-    uint8_t b;
-    uint8_t a;
+    int r;
+    int g;
+    int b;
+    int a;
 };
